requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
PyYAML==6.0.2
pytest==8.3.4
httpx==0.27.2
